+++ eslint.config.js
@@ -2,9 +2,6 @@ const nodePlugin = require('eslint-plugin-n');
 const globals = require('globals');
 
 module.exports = [
-  {
-    ignores: ['test/build/**']
-  },
   nodePlugin.configs['flat/recommended-script'],
   {
     rules: {