+++ src/lib/environment.js
@@ -1,4 +1,4 @@
-const TronBoxError = require('./errors/tronboxerror');
+const TronBoxError = require('./errors/tronBoxError');
 const { expect } = require('./utils');
 const Resolver = require('../components/Resolver');
 const Artifactor = require('../components/Artifactor');
@@ -17,13 +17,6 @@ const Environment = {
       config.artifactor = new Artifactor(config.contracts_build_directory);
     }
 
-    if (!config.network && config.networks['development']) {
-      config.network = 'development';
-    }
-
-    if (!config.network) {
-      return callback(new Error('No network specified. Cannot determine current network.'));
-    }
     const network_config = config.networks[config.network];
 
     if (!network_config) {
@@ -38,7 +31,7 @@ const Environment = {
 
     if (!network_id) {
       return callback(
-        new Error(
+        new TronBoxError(
           "You must specify a network_id in your '" + config.network + "' configuration in order to use this network."
         )
       );