+++ test/fixture/migrations_advanced/1_deploy_contracts.js
@@ -1,0 +1,5 @@
+const Roster = artifacts.require('./Roster.sol');
+
+module.exports = async function (deployer) {
+  await deployer.deploy(Roster, ['Tom', '30'], {});
+};